% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Accordion}
\alias{Accordion}
\title{Accordion}
\usage{
Accordion(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Accordion/ for more information.
}
