% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AccordionActions}
\alias{AccordionActions}
\title{AccordionActions}
\usage{
AccordionActions(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Accordion-Actions/ for more information.
}
