% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AccordionDetails}
\alias{AccordionDetails}
\title{AccordionDetails}
\usage{
AccordionDetails(...)
}
\description{
Visit the documentation at https://mui.com/api/Accordion-Details/ for more information.
}
