% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AccordionSummary}
\alias{AccordionSummary}
\title{AccordionSummary}
\usage{
AccordionSummary(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Accordion-Summary/ for more information.
}
