% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AlertTitle}
\alias{AlertTitle}
\title{AlertTitle}
\usage{
AlertTitle(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Alert-Title/ for more information.
}
