% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AppBar}
\alias{AppBar}
\title{AppBar}
\usage{
AppBar(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/App-Bar/ for more information.
}
