% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{Autocomplete.shinyInput}
\alias{Autocomplete.shinyInput}
\title{Autocomplete}
\usage{
Autocomplete.shinyInput(inputId, ..., value = defaultValue)
}
\description{
Visit the documentation at https://mui.com/api/Autocomplete/ for more information.
}
\examples{
library(shiny.mui)
library(shiny)
library(magrittr)


if (interactive()) {
  shinyApp(
    ui = div(
      Autocomplete.shinyInput("text", options = list(list(label = "go", value = "go"))),
      textOutput("textValue")
    ),
    server = function(input, output) {
      output$textValue <- renderText({
        sprintf("Value: \%s", input$text)
      })
    }
  )
}
}
