% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Avatar}
\alias{Avatar}
\title{Avatar}
\usage{
Avatar(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Avatar/ for more information.
}
