% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AvatarGroup}
\alias{AvatarGroup}
\title{AvatarGroup}
\usage{
AvatarGroup(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Avatar-Group/ for more information.
}
