% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Backdrop}
\alias{Backdrop}
\title{Backdrop}
\usage{
Backdrop(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Backdrop/ for more information.
}
