% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Box}
\alias{Box}
\title{Box}
\usage{
Box(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Box/ for more information.
}
\examples{
library(shiny.mui)
library(shiny)
component_example(tagList(
  Box(backgroundColor = "blue", "Hello world!")
))
}
