% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Breadcrumbs}
\alias{Breadcrumbs}
\title{Breadcrumbs}
\usage{
Breadcrumbs(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Breadcrumbs/ for more information.
}
