% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{button}
\alias{button}
\title{Button
Used to generate button inputs from MUI}
\usage{
button(name)
}
\arguments{
\item{name}{name of js variable}
}
\description{
Button
Used to generate button inputs from MUI
}
