% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ButtonBase}
\alias{ButtonBase}
\title{ButtonBase}
\usage{
ButtonBase(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Button-Base/ for more information.
}
