% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ButtonGroup}
\alias{ButtonGroup}
\title{ButtonGroup}
\usage{
ButtonGroup(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Button-Group/ for more information.
}
