% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Card}
\alias{Card}
\title{Card}
\usage{
Card(...)
}
\description{
Visit the documentation at https://mui.com/api/Card/ for more information.
}
\examples{
library(shiny.mui)
library(shiny)
setInput()
component_example(tagList(
  Card(
    CardContent("Hello world!")
  )
))
}
