% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{CardActionArea}
\alias{CardActionArea}
\title{CardActionArea}
\usage{
CardActionArea(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Card-Action-Area/ for more information.
}
