% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{CardContent}
\alias{CardContent}
\title{CardContent}
\usage{
CardContent(...)
}
\description{
CardContent
}
\examples{
library(shiny.mui)
library(shiny)

component_example(tagList(
  Card(
    CardContent("Hello world!")
  )
))
}
