% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{CardHeader}
\alias{CardHeader}
\title{CardHeader}
\usage{
CardHeader(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Card-Header/ for more information.
}
