% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{CardMedia}
\alias{CardMedia}
\title{CardMedia}
\usage{
CardMedia(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Card-Media/ for more information.
}
