% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Checkbox}
\alias{Checkbox}
\title{Checkbox}
\usage{
Checkbox(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Checkbox/ for more information.
}
