% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{CircularProgress}
\alias{CircularProgress}
\title{CircularProgress}
\usage{
CircularProgress(...)
}
\description{
Visit the documentation at https://mui.com/api/Circular-Progress/ for more information.
}
