% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Collapse}
\alias{Collapse}
\title{Collapse}
\usage{
Collapse(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Collapse/ for more information.
}
