% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{component_example}
\alias{component_example}
\title{Component example}
\usage{
component_example(...)
}
\arguments{
\item{...}{ui taglist to render}
}
\description{
Used for documentation purposes to show shiny examples.
}
