% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Container}
\alias{Container}
\title{Container}
\usage{
Container(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Container/ for more information.
}
