% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{CssBaseline}
\alias{CssBaseline}
\title{CssBaseline}
\usage{
CssBaseline(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Css-Baseline/ for more information.
}
