% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Dialog}
\alias{Dialog}
\title{Dialog}
\usage{
Dialog(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Dialog/ for more information.
}
