% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{DialogActions}
\alias{DialogActions}
\title{DialogActions}
\usage{
DialogActions(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Dialog-Actions/ for more information.
}
