% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{DialogContentText}
\alias{DialogContentText}
\title{DialogContentText}
\usage{
DialogContentText(...)
}
\description{
Visit the documentation at https://mui.com/api/Dialog-Content-Text/ for more information.
}
