% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{DialogTitle}
\alias{DialogTitle}
\title{DialogTitle}
\usage{
DialogTitle(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Dialog-Title/ for more information.
}
