% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Divider}
\alias{Divider}
\title{Divider}
\usage{
Divider(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Divider/ for more information.
}
