% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Drawer}
\alias{Drawer}
\title{Drawer}
\usage{
Drawer(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Drawer/ for more information.
}
