% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{FilledInput}
\alias{FilledInput}
\title{FilledInput}
\usage{
FilledInput(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Filled-Input/ for more information.
}
