% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{FormControl}
\alias{FormControl}
\title{FormControl}
\usage{
FormControl(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Form-Control/ for more information.
}
