% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{FormControlLabel}
\alias{FormControlLabel}
\title{FormControlLabel}
\usage{
FormControlLabel(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Form-Control-Label/ for more information.
}
