% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{FormGroup}
\alias{FormGroup}
\title{FormGroup}
\usage{
FormGroup(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Form-Group/ for more information.
}
