% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{GlobalStyles}
\alias{GlobalStyles}
\title{GlobalStyles}
\usage{
GlobalStyles(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Global-Styles/ for more information.
}
