% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Hidden}
\alias{Hidden}
\title{Hidden}
\usage{
Hidden(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Hidden/ for more information.
}
