% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icon}
\alias{icon}
\title{Icon
Used to import Icons from Mui}
\usage{
icon(name, module = "@mui/icons-material")
}
\arguments{
\item{name}{name of js variable}

\item{module}{js module exported from js package where component is defined}
}
\description{
Icon
Used to import Icons from Mui
}
