% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{IconButton}
\alias{IconButton}
\title{IconButton}
\usage{
IconButton(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Icon-Button/ for more information.
}
