% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ImageList}
\alias{ImageList}
\title{ImageList}
\usage{
ImageList(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Image-List/ for more information.
}
