% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ImageListItem}
\alias{ImageListItem}
\title{ImageListItem}
\usage{
ImageListItem(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Image-List-Item/ for more information.
}
