% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ImageListItemBar}
\alias{ImageListItemBar}
\title{ImageListItemBar}
\usage{
ImageListItemBar(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Image-List-Item-Bar/ for more information.
}
