% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{input}
\alias{input}
\title{Input
Used to generate inputs from MUI}
\usage{
input(name, defaultValue = "")
}
\arguments{
\item{name}{name of js variable}

\item{defaultValue}{the default value for the element}
}
\description{
Input
Used to generate inputs from MUI
}
