% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{InputAdornment}
\alias{InputAdornment}
\title{InputAdornment}
\usage{
InputAdornment(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Input-Adornment/ for more information.
}
