% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{InputBase}
\alias{InputBase}
\title{InputBase}
\usage{
InputBase(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Input-Base/ for more information.
}
