% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{InputLabel}
\alias{InputLabel}
\title{InputLabel}
\usage{
InputLabel(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Input-Label/ for more information.
}
