% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{LinearProgress}
\alias{LinearProgress}
\title{LinearProgress}
\usage{
LinearProgress(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Linear-Progress/ for more information.
}
