% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{List}
\alias{List}
\title{List}
\usage{
List(...)
}
\description{
List
}
\examples{
library(shiny)
library(shiny.mui)


component_example(List(
    ListItem(ListItemText(primary = "Hello", secondary = "First")),
    ListItem(ListItemText(primary = "World", secondary = "second")),
))
}
