% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ListItem}
\alias{ListItem}
\title{ListItem}
\usage{
ListItem(...)
}
\description{
ListItem
}
\examples{
library(shiny)
library(shiny.mui)


component_example(List(
    ListItem(ListItemText(primary = "Hello", secondary = "First")),
    ListItem(ListItemText(primary = "World", secondary = "second")),
))
}
