% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ListItemAvatar}
\alias{ListItemAvatar}
\title{ListItemAvatar}
\usage{
ListItemAvatar(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/List-Item-Avatar/ for more information.
}
