% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ListItemButton}
\alias{ListItemButton}
\title{ListItemButton}
\usage{
ListItemButton(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/List-Item-Button/ for more information.
}
