% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ListItemIcon}
\alias{ListItemIcon}
\title{ListItemIcon}
\usage{
ListItemIcon(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/List-Item-Icon/ for more information.
}
