% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ListItemSecondaryAction}
\alias{ListItemSecondaryAction}
\title{ListItemSecondaryAction}
\usage{
ListItemSecondaryAction(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/List-Item-Secondary-Action/ for more information.
}
