% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ListSubheader}
\alias{ListSubheader}
\title{ListSubheader}
\usage{
ListSubheader(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/List-Subheader/ for more information.
}
