% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{MenuItem}
\alias{MenuItem}
\title{MenuItem}
\usage{
MenuItem(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Menu-Item/ for more information.
}
