% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{MenuList}
\alias{MenuList}
\title{MenuList}
\usage{
MenuList(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Menu-List/ for more information.
}
