% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{MobileStepper}
\alias{MobileStepper}
\title{MobileStepper}
\usage{
MobileStepper(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Mobile-Stepper/ for more information.
}
