% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{OutlinedInput}
\alias{OutlinedInput}
\title{OutlinedInput}
\usage{
OutlinedInput(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Outlined-Input/ for more information.
}
