% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Pagination}
\alias{Pagination}
\title{Pagination}
\usage{
Pagination(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Pagination/ for more information.
}
