% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{PaginationItem}
\alias{PaginationItem}
\title{PaginationItem}
\usage{
PaginationItem(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Pagination-Item/ for more information.
}
