% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Popper}
\alias{Popper}
\title{Popper}
\usage{
Popper(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Popper/ for more information.
}
