% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Portal}
\alias{Portal}
\title{Portal}
\usage{
Portal(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Portal/ for more information.
}
