% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Rating}
\alias{Rating}
\title{Rating}
\usage{
Rating(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Rating/ for more information.
}
