% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{Select.shinyInput}
\alias{Select.shinyInput}
\title{Select}
\usage{
Select.shinyInput(inputId, ..., value = defaultValue)
}
\description{
Select
}
\examples{
library(shiny.mui)
library(shiny)

if (interactive()) {
  shinyApp(
    ui = div(
      TextField.shinyInput("text", value = "go"),
      textOutput("textValue")
    ),
    server = function(input, output) {
      output$textValue <- renderText({
        sprintf("Value: \%s", input$text)
      })
    }
  )
}
}
