% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Skeleton}
\alias{Skeleton}
\title{Skeleton}
\usage{
Skeleton(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Skeleton/ for more information.
}
