% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Slider}
\alias{Slider}
\title{Slider}
\usage{
Slider(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Slider/ for more information.
}
