% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{Slider.shinyInput}
\alias{Slider.shinyInput}
\title{Slider}
\usage{
Slider.shinyInput(inputId, ..., value = defaultValue)
}
\description{
Visit the documentation at https://mui.com/api/Slider/ for more information.
}
\examples{
library(shiny.mui)
library(shiny)
library(magrittr)


if (interactive()) {
  shinyApp(
    ui = div(
      Slider.shinyInput("text", value = 5),
      textOutput("textValue")
    ),
    server = function(input, output) {
      output$textValue <- renderText({
        sprintf("Value: \%s", input$text)
      })
    }
  )
}
}
