% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Snackbar}
\alias{Snackbar}
\title{Snackbar}
\usage{
Snackbar(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Snackbar/ for more information.
}
