% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{SnackbarContent}
\alias{SnackbarContent}
\title{SnackbarContent}
\usage{
SnackbarContent(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Snackbar-Content/ for more information.
}
