% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{SpeedDialAction}
\alias{SpeedDialAction}
\title{SpeedDialAction}
\usage{
SpeedDialAction(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Speed-Dial-Action/ for more information.
}
