% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{SpeedDialIcon}
\alias{SpeedDialIcon}
\title{SpeedDialIcon}
\usage{
SpeedDialIcon(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Speed-Dial-Icon/ for more information.
}
