% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{StepButton}
\alias{StepButton}
\title{StepButton}
\usage{
StepButton(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Step-Button/ for more information.
}
