% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{StepConnector}
\alias{StepConnector}
\title{StepConnector}
\usage{
StepConnector(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Step-Connector/ for more information.
}
