% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{StepContent}
\alias{StepContent}
\title{StepContent}
\usage{
StepContent(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Step-Content/ for more information.
}
