% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{StepIcon}
\alias{StepIcon}
\title{StepIcon}
\usage{
StepIcon(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Step-Icon/ for more information.
}
