% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{StepLabel}
\alias{StepLabel}
\title{StepLabel}
\usage{
StepLabel(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Step-Label/ for more information.
}
