% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{StyledEngineProvider}
\alias{StyledEngineProvider}
\title{StyledEngineProvider}
\usage{
StyledEngineProvider(...)
}
\description{
Visit the documentation at https://mui.com/api/Styled-Engine-Provider/ for more information.
}
