% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{SwipeableDrawer}
\alias{SwipeableDrawer}
\title{SwipeableDrawer}
\usage{
SwipeableDrawer(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Swipeable-Drawer/ for more information.
}
