% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{Switch.shinyInput}
\alias{Switch.shinyInput}
\title{Select}
\usage{
Switch.shinyInput(inputId, ..., value = defaultValue)
}
\description{
Visit the documentation at https://mui.com/api/Switch/ for more information.
}
\examples{
library(shiny.mui)
library(shiny)
library(magrittr)


if (interactive()) {
  shinyApp(
    ui = div(
      Switch.shinyInput("text", value = T),
      textOutput("textValue")
    ),
    server = function(input, output) {
      output$textValue <- renderText({
        sprintf("Value: \%s", input$text)
      })
    }
  )
}
}
