% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{TableBody}
\alias{TableBody}
\title{TableBody}
\usage{
TableBody(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Table-Body/ for more information.
}
