% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{TableCell}
\alias{TableCell}
\title{TableCell}
\usage{
TableCell(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Table-Cell/ for more information.
}
