% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{TableContainer}
\alias{TableContainer}
\title{TableContainer}
\usage{
TableContainer(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Table-Container/ for more information.
}
