% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{TableFooter}
\alias{TableFooter}
\title{TableFooter}
\usage{
TableFooter(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Table-Footer/ for more information.
}
