% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{TableSortLabel}
\alias{TableSortLabel}
\title{TableSortLabel}
\usage{
TableSortLabel(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Table-Sort-Label/ for more information.
}
