% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{Tabs.shinyInput}
\alias{Tabs.shinyInput}
\title{Tabs}
\usage{
Tabs.shinyInput(inputId, ..., value = defaultValue)
}
\description{
Visit the documentation at https://mui.com/api/Tabs/ for more information.
}
\examples{
library(shiny.mui)
library(shiny)

if (interactive()) {
  shinyApp(
    ui = div(
      Tabs.shinyInput("tabs", value = "first", Tab(value = "first", label = "First", "First"), Tab(value = "second", label = "Second", "Second"), Tab(value = "third", label = "Third", "Third")),
      textOutput("textValue")
    ),
    server = function(input, output) {
      output$textValue <- renderText({
        sprintf("Value: \%s", input$tabs)
      })
    }
  )
}
}
