% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{TextareaAutosize}
\alias{TextareaAutosize}
\title{TextareaAutosize}
\usage{
TextareaAutosize(...)
}
\description{
Visit the documentation at https://mui.com/api/Textarea-Autosize/ for more information.
}
