% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ThemeProvider}
\alias{ThemeProvider}
\title{ThemeProvider}
\usage{
ThemeProvider(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Theme-Provider/ for more information.
}
