% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ToggleButtonGroup}
\alias{ToggleButtonGroup}
\title{ToggleButtonGroup}
\usage{
ToggleButtonGroup(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Toggle-Button-Group/ for more information.
}
