% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Toolbar}
\alias{Toolbar}
\title{Toolbar}
\usage{
Toolbar(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Toolbar/ for more information.
}
