% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Tooltip}
\alias{Tooltip}
\title{Tooltip}
\usage{
Tooltip(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Tooltip/ for more information.
}
