% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Typography}
\alias{Typography}
\title{Typography}
\usage{
Typography(...)
}
\description{
Typography
}
\examples{
library(shiny.mui)
library(shiny)

component_example(tagList(
  Typography(variant="h1", "H1"),
  Typography(variant="h2", "H2"),
  Typography(variant="body1", "Body 1"),
))
}
