% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{BottomNavigation}
\alias{BottomNavigation}
\title{BottomNavigation}
\usage{
BottomNavigation(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Bottom-Navigation/ for more information.
}
