% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{BottomNavigationAction}
\alias{BottomNavigationAction}
\title{BottomNavigationAction}
\usage{
BottomNavigationAction(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Bottom-Navigation-Action/ for more information.
}
