% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{Button.shinyInput}
\alias{Button.shinyInput}
\title{Button}
\usage{
Button.shinyInput(inputId, ...)
}
\description{
Button
}
\examples{
library(shiny.mui)
library(shiny)
library(magrittr)

component_example(tagList(
  Button.shinyInput("button1", "Text", variant="text"),
  Button.shinyInput("button1", "Contained", variant="contained"),
  Button.shinyInput("button1", "Outlined", variant="outlined"),
))
}
