% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{CardContent}
\alias{CardContent}
\title{CardContent}
\usage{
CardContent(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Card-Content/ for more information.
}
\examples{
library(shiny.mui)
library(shiny)
component_example(tagList(
  Card(
    CardContent("Hello world!")
  )
))
}
