% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ClickAwayListener}
\alias{ClickAwayListener}
\title{ClickAwayListener}
\usage{
ClickAwayListener(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Click-Away-Listener/ for more information.
}
