% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{IconButton.shinyInput}
\alias{IconButton.shinyInput}
\title{IconButton}
\usage{
IconButton.shinyInput(inputId, ...)
}
\arguments{
\item{inputId}{the shiny input id}

\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Icon-Button/ for more information.
}
