% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ListItemText}
\alias{ListItemText}
\title{ListItemText}
\usage{
ListItemText(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/List-Item-Text/ for more information.
}
\examples{
library(shiny)
library(shiny.mui)


component_example(List(
    ListItem(ListItemText(primary = "Hello", secondary = "First")),
    ListItem(ListItemText(primary = "World", secondary = "second")),
))
}
