% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ScopedCssBaseline}
\alias{ScopedCssBaseline}
\title{ScopedCssBaseline}
\usage{
ScopedCssBaseline(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Scoped-Css-Baseline/ for more information.
}
