% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{Slider.shinyInput}
\alias{Slider.shinyInput}
\title{Slider}
\usage{
Slider.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{inputId}{the shiny input id}

\item{...}{args to pass to element}

\item{value}{the default value of the element
Visit the documentation at https://mui.com/api/Slider/ for more information.}
}
\description{
Slider
}
\examples{
library(shiny.mui)
library(shiny)



if (interactive()) {
  shinyApp(
    ui = div(
      Slider.shinyInput("text", value = 5),
      textOutput("textValue")
    ),
    server = function(input, output) {
      output$textValue <- renderText({
        sprintf("Value: \%s", input$text)
      })
    }
  )
}
}
