% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{Tabs.shinyInput}
\alias{Tabs.shinyInput}
\title{Tabs}
\usage{
Tabs.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{inputId}{the shiny input id}

\item{...}{args to pass to element}

\item{value}{the default value of the element}
}
\description{
Visit the documentation at https://mui.com/api/Tabs/ for more information.
}
\examples{
library(shiny.mui)
library(shiny)

if (interactive()) {
  shinyApp(
    ui = div(
      Tabs.shinyInput(
        "tabs",
        value = "first",
        Tab(value = "first", label = "First"),
        Tab(value = "second", label = "Second"),
        Tab(value = "third", label = "Third")
      ),
      textOutput("text_value")
    ),
    server = function(input, output) {
      output$text_value <- renderText({
        glue:glue("Value: {input$tabs}")
      })
    }
  )
}

}
