% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{TextField.shinyInput}
\alias{TextField.shinyInput}
\title{TextField}
\usage{
TextField.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{inputId}{the shiny input id}

\item{...}{args to pass to element}

\item{value}{the default value of the element}
}
\description{
Visit the documentation at https://mui.com/api/Text-Field/ for more information.
}
\examples{
library(shiny.mui)
library(shiny)

if (interactive()) {
  shinyApp(
    ui = div(
      TextField.shinyInput("text", fullWidth = T, value = "go", variant="standard", multiline = T),
      textOutput("textValue")
    ),
    server = function(input, output) {
      output$textValue <- renderText({
        sprintf("Value: \%s", input$text)
      })
    }
  )
}
}
